+++ common/src/main/java/ctbrec/Model.java
@@ -9,6 +9,7 @@ import com.iheartradio.m3u8.PlaylistException;
 import com.squareup.moshi.JsonReader;
 import com.squareup.moshi.JsonWriter;
 
+import ctbrec.recorder.download.Download;
 import ctbrec.recorder.download.StreamSource;
 import ctbrec.sites.Site;
 
@@ -101,6 +102,6 @@ public interface Model extends Comparable<Model> {
 
     public void setSuspended(boolean suspended);
 
-
+    public Download createDownload();
 
 }